Move gap summary mock to the common api folder
